VIITE-1603 Move ApplyChangeInfoProcess to new file
